// learning-app-fe/src/enums/LessonLevel.ts
export enum LessonLevel {
  N5_BEGINNER = "N5_BEGINNER", // N5 sơ cấp
  N5_ELEMENTARY = "N5_ELEMENTARY", // N5 sơ trung cấp

  N4_BEGINNER = "N4_BEGINNER", // N4 sơ cấp
  N4_ELEMENTARY = "N4_ELEMENTARY", // N4 sơ trung cấp

  N3_INTERMEDIATE = "N3_INTERMEDIATE", // N3 trung cấp

  N2_UPPER = "N2_UPPER", // N2 trung – cao cấp
  N1_ADVANCED = "N1_ADVANCED", // N1 cao cấp
}